Refactor Rating model
This commit moves the `Rating` data class from `SelfRatingsResponseJson.kt` to its own file `Rating.kt` under `com.saintpatrck.mealie.client.api.model`.

This change improves code organization by separating the model from the response DTO.
